+++ src/Service/MediaConvert/src/Enum/AacCodecProfile.php
@@ -3,7 +3,10 @@
 namespace AsyncAws\MediaConvert\Enum;
 
 /**
- * AAC Profile.
+ * Specify the AAC profile. For the widest player compatibility and where higher bitrates are acceptable: Keep the
+ * default profile, LC (AAC-LC) For improved audio performance at lower bitrates: Choose HEV1 or HEV2. HEV1 (AAC-HE v1)
+ * adds spectral band replication to improve speech audio at low bitrates. HEV2 (AAC-HE v2) adds parametric stereo,
+ * which optimizes for encoding stereo audio at very low bitrates.
  */
 final class AacCodecProfile
 {