+++ src/Service/MediaConvert/src/ValueObject/H264Settings.php
@@ -20,6 +20,7 @@ use AsyncAws\MediaConvert\Enum\H264ParControl;
 use AsyncAws\MediaConvert\Enum\H264QualityTuningLevel;
 use AsyncAws\MediaConvert\Enum\H264RateControlMode;
 use AsyncAws\MediaConvert\Enum\H264RepeatPps;
+use AsyncAws\MediaConvert\Enum\H264SaliencyAwareEncoding;
 use AsyncAws\MediaConvert\Enum\H264ScanTypeConversionMode;
 use AsyncAws\MediaConvert\Enum\H264SceneChangeDetect;
 use AsyncAws\MediaConvert\Enum\H264SlowPal;
@@ -264,16 +265,17 @@ final class H264Settings
     private $maxBitrate;
 
     /**
-     * Use this setting only when you also enable Scene change detection. This setting determines how the encoder manages
-     * the spacing between I-frames that it inserts as part of the I-frame cadence and the I-frames that it inserts for
-     * Scene change detection. We recommend that you have the transcoder automatically choose this value for you based on
-     * characteristics of your input video. To enable this automatic behavior, do this by keeping the default empty value.
-     * When you explicitly specify a value for this setting, the encoder determines whether to skip a cadence-driven I-frame
-     * by the value you set. For example, if you set Min I interval to 5 and a cadence-driven I-frame would fall within 5
-     * frames of a scene-change I-frame, then the encoder skips the cadence-driven I-frame. In this way, one GOP is shrunk
-     * slightly and one GOP is stretched slightly. When the cadence-driven I-frames are farther from the scene-change
-     * I-frame than the value you set, then the encoder leaves all I-frames in place and the GOPs surrounding the scene
-     * change are smaller than the usual cadence GOPs.
+     * Specify the minimum number of frames allowed between two IDR-frames in your output. This includes frames created at
+     * the start of a GOP or a scene change. Use Min I-Interval to improve video compression by varying GOP size when two
+     * IDR-frames would be created near each other. For example, if a regular cadence-driven IDR-frame would fall within 5
+     * frames of a scene-change IDR-frame, and you set Min I-interval to 5, then the encoder would only write an IDR-frame
+     * for the scene-change. In this way, one GOP is shortened or extended. If a cadence-driven IDR-frame would be further
+     * than 5 frames from a scene-change IDR-frame, then the encoder leaves all IDR-frames in place. To use an automatically
+     * determined interval: We recommend that you keep this value blank. This allows for MediaConvert to use an optimal
+     * setting according to the characteristics of your input video, and results in better video compression. To manually
+     * specify an interval: Enter a value from 1 to 30. Use when your downstream systems have specific GOP size
+     * requirements. To disable GOP size variance: Enter 0. MediaConvert will only create IDR-frames at the start of your
+     * output's cadence-driven GOP. Use when your downstream systems require a regular GOP size.
      *
      * @var int|null
      */
@@ -359,6 +361,18 @@ final class H264Settings
      * @var H264RepeatPps::*|null
      */
     private $repeatPps;
+
+    /**
+     * Specify whether to apply Saliency aware encoding to your output. Use to improve the perceptual video quality of your
+     * output by allocating more encoding bits to the prominent or noticeable parts of your content. To apply saliency aware
+     * encoding, when possible: We recommend that you choose Preferred. The effects of Saliency aware encoding are best seen
+     * in lower bitrate outputs. When you choose Preferred, note that Saliency aware encoding will only apply to outputs
+     * that are 720p or higher in resolution. To not apply saliency aware encoding, prioritizing encoding speed over
+     * perceptual video quality: Choose Disabled.
+     *
+     * @var H264SaliencyAwareEncoding::*|null
+     */
+    private $saliencyAwareEncoding;
 
     /**
      * Use this setting for interlaced outputs, when your output frame rate is half of your input frame rate. In this
@@ -515,6 +529,7 @@ final class H264Settings
      *   QvbrSettings?: null|H264QvbrSettings|array,
      *   RateControlMode?: null|H264RateControlMode::*,
      *   RepeatPps?: null|H264RepeatPps::*,
+     *   SaliencyAwareEncoding?: null|H264SaliencyAwareEncoding::*,
      *   ScanTypeConversionMode?: null|H264ScanTypeConversionMode::*,
      *   SceneChangeDetect?: null|H264SceneChangeDetect::*,
      *   Slices?: null|int,
@@ -562,6 +577,7 @@ final class H264Settings
         $this->qvbrSettings = isset($input['QvbrSettings']) ? H264QvbrSettings::create($input['QvbrSettings']) : null;
         $this->rateControlMode = $input['RateControlMode'] ?? null;
         $this->repeatPps = $input['RepeatPps'] ?? null;
+        $this->saliencyAwareEncoding = $input['SaliencyAwareEncoding'] ?? null;
         $this->scanTypeConversionMode = $input['ScanTypeConversionMode'] ?? null;
         $this->sceneChangeDetect = $input['SceneChangeDetect'] ?? null;
         $this->slices = $input['Slices'] ?? null;
@@ -609,6 +625,7 @@ final class H264Settings
      *   QvbrSettings?: null|H264QvbrSettings|array,
      *   RateControlMode?: null|H264RateControlMode::*,
      *   RepeatPps?: null|H264RepeatPps::*,
+     *   SaliencyAwareEncoding?: null|H264SaliencyAwareEncoding::*,
      *   ScanTypeConversionMode?: null|H264ScanTypeConversionMode::*,
      *   SceneChangeDetect?: null|H264SceneChangeDetect::*,
      *   Slices?: null|int,
@@ -843,6 +860,14 @@ final class H264Settings
     }
 
     /**
+     * @return H264SaliencyAwareEncoding::*|null
+     */
+    public function getSaliencyAwareEncoding(): ?string
+    {
+        return $this->saliencyAwareEncoding;
+    }
+
+    /**
      * @return H264ScanTypeConversionMode::*|null
      */
     public function getScanTypeConversionMode(): ?string
@@ -1071,6 +1096,12 @@ final class H264Settings
                 throw new InvalidArgument(\sprintf('Invalid parameter "repeatPps" for "%s". The value "%s" is not a valid "H264RepeatPps".', __CLASS__, $v));
             }
             $payload['repeatPps'] = $v;
+        }
+        if (null !== $v = $this->saliencyAwareEncoding) {
+            if (!H264SaliencyAwareEncoding::exists($v)) {
+                throw new InvalidArgument(\sprintf('Invalid parameter "saliencyAwareEncoding" for "%s". The value "%s" is not a valid "H264SaliencyAwareEncoding".', __CLASS__, $v));
+            }
+            $payload['saliencyAwareEncoding'] = $v;
         }
         if (null !== $v = $this->scanTypeConversionMode) {
             if (!H264ScanTypeConversionMode::exists($v)) {